# backend/docs/swagger.yaml
basePath: /
definitions:
  internal_httpserver_handlers.ErrorResponse:
    properties:
      error:
        type: string
      ok:
        type: boolean
    type: object
  internal_httpserver_handlers.GenericObject:
    additionalProperties: {}
    type: object
...
  internal_httpserver_handlers.NewsIngestJSONBody:
    properties:
      payload:
        additionalProperties: {}
        type: object
      topic:
        type: string
    type: object
  internal_httpserver_handlers.OkResponse:
    properties:
      ok:
        type: boolean
    type: object
...
info:
  contact: {}
  description: 后端基于 Gin 的 HTTP API 文档（swaggo/swag 自动生成）。
  title: F1InkDashboard Backend API
  version: 1.0.0
paths:
...
  /api/v1/mp/news:
    get:
      description: 分页返回新闻条目列表（不包含 content）。
      parameters:
      - default: Asia/Shanghai
        description: IANA 时区名称
        in: query
        name: tz
        type: string
      - default: 1
        description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量（最大 50）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.GenericObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      summary: 小程序新闻列表
      tags:
      - MiniProgram
  /api/v1/mp/news/{id}:
    get:
      description: 返回指定 id 的新闻详情（包含 content）。
      parameters:
      - description: 新闻 ID
        in: path
        name: id
        required: true
        type: string
      - default: Asia/Shanghai
        description: IANA 时区名称
        in: query
        name: tz
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.GenericObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      summary: 小程序新闻详情
      tags:
      - MiniProgram
...
  /api/v1/mp/telemetry/sector_controls:
    get:
      description: 返回 points[{x, sector, t_ms, speed, throttle, brake}]。
      parameters:
      - description: OpenF1 session_key
        in: query
        name: session_key
        required: true
        type: integer
      - description: 车手号码
        in: query
        name: driver_number
        required: true
        type: integer
      - default: 0
        description: 最大采样点（0 表示默认策略，最大 20000）
        in: query
        name: max_points
        type: integer
      - default: fastest
        description: 选择圈：fastest/all 等
        in: query
        name: lap
        type: string
      - description: 指定圈号
        in: query
        name: lap_number
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.GenericObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      summary: 分段油门/刹车曲线（带 sector 归属）
      tags:
      - MiniProgram
  /api/v1/news/ingest:
    post:
      consumes:
      - application/json
      description: '|'
      parameters:
      - description: 鉴权 token（当 NEWS_INGEST_TOKEN 非空时必填）
        in: query
        name: token
        type: string
      - description: 注入内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpserver_handlers.NewsIngestJSONBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      security:
      - TokenQuery: []
      summary: 注入新闻（JSON）
      tags:
      - News
  /api/v1/news/meme/ws/ingest:
    post:
      consumes:
      - multipart/form-data
      description: '|'
      parameters:
      - description: 鉴权 token（当 NEWS_INGEST_TOKEN 非空时必填）
        in: query
        name: token
        type: string
      - description: 标题
        in: formData
        name: title
        required: true
        type: string
      - description: 可选图片文件
        in: formData
        name: image
        type: file
      - description: 可选音频文件
        in: formData
        name: audio
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      security:
      - TokenQuery: []
      summary: 注入 Meme 新闻（multipart）
      tags:
      - News
  /api/v1/news/ws/ingest:
    post:
      consumes:
      - multipart/form-data
      description: '|'
      parameters:
      - description: 鉴权 token（当 NEWS_INGEST_TOKEN 非空时必填）
        in: query
        name: token
        type: string
      - description: 新闻标题
        in: formData
        name: title
        required: true
        type: string
      - description: 可选图片文件
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      security:
      - TokenQuery: []
      summary: 注入 Breaking 新闻（multipart）
      tags:
      - News
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Authorization: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
  TokenQuery:
    description: 简单鉴权 token（NEWS_INGEST_TOKEN / OPENF1_INGEST_TOKEN / WECHATPAY_API_TOKEN
      等）
    in: query
    name: token
    type: apiKey
swagger: "2.0"